Fix incomplete PHPStorm refactor and move setConfigSingleton event handler in TestingEnvironment to TestConfig object.
