+++ tests/PHPUnit/proxy/index.php
@@ -6,6 +6,6 @@
 
 require realpath(dirname(__FILE__)) . "/includes.php";
 
-Piwik_TestingEnvironment::addHooks();
+\Piwik\Tests\Framework\TestingEnvironment::addHooks();
 
 include PIWIK_INCLUDE_PATH . '/index.php';
